[Runtime] Decrease initialization priority of `gc` (#52)
A recent change in MSVCs debug library is incompatible with the very low
priority we set causing all executables to crash on exit. This happens
due to initialization and deinitialization/destruction order causing
undefined behaviour as objects are accessed in unsafe manners.

We will likely want to consider initializing `gc` and others in `main`
in the future, to avoid any such issues.