[ci] Enable merge queue (#56)
Having a merge queue allows us to drop the requirement for a branch to
be up-to-date to merge a PR. Rather, if it has no conflicts, it gets
added to the queue and automatically tested on top of `main`. This
automates things a lot as we can just push PRs to the queue and wait for
it to get merged together with other PRs.