[Py] Introduce `IntAttr`- and `BoolAttrInterface`
These interfaces can be implemented by any attribute that subclasses or in some way "is" an integer or boolean respectively.
Compared to the current approach, this is much more powerful as it does not dictate the concrete storage format used by an implementation.
A `GlobalValueAttr` or a hypothetical `DerivedAttr` would therefore be able to easily implement these interfaces.