[Py] Split `ConstObjectAttrInterface` from `ObjectAttrInterface` (#59)
The primary motivation for this change is that `getSlots` is only valid
if we know that the python object being represented by the attribute is
a constant. This is most notably not the case for `#py.globalValue` or
`#py.ref` that isn't marked `const` or is not initialized with a
constant object.