Rework clang-format and clang-tidy style
This commit changes the formatting to a style that I organically grew over time to something more akin to LLVMs style.
The reason for this change is that through my work in upstream LLVM and MLIR I've become too accustomed to their codestyle.
More objectively, any contributors to Pylir are also more likely to be accustomed to MLIRs codestyle rather than whatever I made up myself.
Doing this change earlier than later makes it easier and possible to adjust the whole codebase as well.