[ci] Add nightly recompressing ccache cache
CCache uses `zstd` for compression and decompression which has the nice property of having the same decompression speed regardless of compression level used. To save on cache space and therefore improve cache hit rate, it is potentially profitable to recompress the cache every night with the maximum compression level possible.