[Py] Change `getTypeObject` to return `TypeAttrInterface` (#75)
There is no reason as to why it must be a `RefAttr` specifically, and
with the new casting ability of `RefAttr`, it is possible to use
`TypeAttrInterface` instead. The only disadvantage is the explicit cast
required which may possibly defer an error to use, rather than
construction, which is less ergonomic. This can be mitigated by future
implementations of attribute verifiers and `GlobalValueAttr`.