[Py] Remove `type_object` and `slots` from most attributes (#76)
The rationale for this change has multiple facets:
* It reduces the complexity of the implementations of each attributes as
they now represent a single concept "A builtin instance of this python
object" with less state.
* It makes it possible in the future we can relax the requirement of
builtins being defined, making it possibly a phase invariant rather than
an IR invariant instead (as arguably, the type object only needs to be
known for lowering and optimization).
* There is now tiny less magic in regards to a "hidden-reference to a
type object" as it now ony a "hidden-reference to a builtin type object"
* Since `RefAttr`s are no longer part of the interfaces of these types,
it should make porting to `GlobalValueAttr` require less changes.

This PR is technically speaking not NFC as it removes functionality.
However, using a different type object than a builtin was untested and
unused and I would therefore consider tech debt. This functionality can
be restored in a likely cleaner way by creating a `DerivedAttr` in the
future.