[CMake] Add and default to unified Pylir and LLVM build
The unified build makes it possible to build the required parts of LLVM with the exact revision required as part of the normal Pylir build.
This now creates a better out of the box experience of compiling the project by only requiring a single build step instead of separately compiling LLVM.

Furthermore, this has the potential to greatly improve developer experience as bumping an LLVM version is now automatic, both performing it oneself and receiving a commit that bumped LLVM.
Other advantages include the possibility to work on LLVM and Pylir at the same time and improved total build times due to finer grained target dependency tracking.

Note that the main CI builders will not be using this kind of build as it makes caching just the LLVM build probably more difficult and at the very least incompatible with the current mechanism.