[Py] Simplify `DictAttr` (#85)
The previous implementation relied on complex machinery for finding the
kind of builtin hash function used for a given key attribute. This has
become mostly redundant as most attributes no longer have a type object
that could modify the hashing function.

The new implementation is based on a new `EqualsInterface` which returns
the canonical attribute, improving the organization. Futhermore, for the
time being, the capability of using a `RefAttr` as key was removed until
deemed necessary. This can be achieved in the future by once again
introducing the explicit normalized key attribute in the syntax.
Similarly, support could be added for all python attributes that use
identity hash/comparison as well. This should be added on a per-need
basis.