[Py] Replace `RefAttr` and `GlobaValueOp` in favour of `GlobalValueAttr` (#86)
`RefAttr` and `GlobalValueOp`s were one of the more problematic
attributes in the compiler with many pitfalls such as the need to link
them prior to verification, the odd split of both an "op" + a "ref",
which caused other issues such as the inability to create a new
`GlobalValueOp` in a multi-threaded context.

This PR replaces the two with the previously introduced
`GlobalValueAttr`. While not perfect, it removes the "op" + "ref attr"
split by being just a single attribute that can be used directly. It
allows us to remove the `FinalizeRefAttr` pass and custom `mlir-opt`
implementation in favour of the upstream implementation and gives us
potential for many future improvements to IR structure.