[test] Add cmake option for execution test compile flags (#87)
The execution tests currently serve as 1) tests for the correct
implementation of python builtins and 2) as e2e tests, acting as our
last line of defense. For the purpose of 2) it is useful to run the
tests with flags, such as higher optimization levels. This PR adds a
mechanism to do so by adding the cmake flag `PYLIR_TEST_EXEC_FLAGS`
which is added to any invocation of `pylir` in the `test/Execution`
directory.