[Py] Add attribute verifier to verify structural properties (#88)
Few of our attributes tend to require verifying structural traits that
cannot be verified on construction. Previously, verification of these
were hardcoded parts of the op verification which was not nicely
constructed and made use of a large `TypeSwitch`.

The new implementation makes use of an attribute interface and an
external driver that can be attached to operations via a trait. The
former allows attributes to opt into verification and delegates the
verification behaviour to the attribute itself rather than the large
central `TypeSwitch` from before. The external driver also takes care of
the recursive traversal of attributes which had previously been
implemented manually in the op verifiers.