[Py] Introduce `ObjectBaseAttribute` and `ConcreteObjectAttribute` base classes
Previously, interfaces such as `ObjectAttrInterface` or `ConcreteObjectInterface` were being used as "constraints" in methods and even as parameters of attributes despite their interface methods never to rarely being used. As interfaces are twice as large as plain attributes and converting to the interface requires a hashmap lookup in the context, using the interface as constraint in many of these scenarios cost us both in time, space and arguably comprehension.
Furthermore, `ObjectAttrInterface` is only conditionally implemented by `GlobalValueAttr` and therefore not suitable in almost all scenarios currently used.

This commit therefore introduces base classes implemented by the attributes in `PylirPyAttributes.td` that act as a more constrained version of `mlir::Attribute`.