[Py] Make `ObjectAttrInterface`'s `getTypeObject` return `ObjectBaseAttribute`
This relaxes the requirements imposed by any attribute implementing the interface. It is no longer required by the `getTypeObject` implementation to return an implementation of `TypeAttrInterface`, but rather any kind of object attribute. The motivation of this relaxation is to allow `#py.obj`, future derived attributes and most importantly the current concrete attributes such as `#py.int` to return `#py.globalValue` instances that do not yet have an initializer.
For the majority of the compilers passes, the type object being defined is a pure optimization. Future changes will attempt to either document or drop this requirement from the LLVM lowering as well.