[PyHIRToPylirPy] Initial Lowering of `pyHIR.init` to `py` (#98)
While the two dialects are largely meant to live side-by-side throughout
most of the compiler, there will be later phases requiring all `pyHIR`
operations to have disappeared for further lowering and optimizations.
This PR is the first step in lowering the dialect and working towards an
E2E flow with the `pyHIR` dialect and `CodeGenNew` Frontend lowering.