[Main] Remove support for ASAN instrumentation (#101)
Similar to
https://github.com/Pylir/Pylir/commit/8700913d05d785262eec2376c1b74041e9a4f23f,
using ASAN with address spaces never really worked and with recent
changes to upstream even causes assertion failures. Until it is
officially supported, remove ASAN codegen capabilities. Note that
linking against TSAN and ASAN is still supported and used when the C++
runtime is built with either of these instrumentations.