[pyHIR] Add `func` outlining pass (#103)
This pass can be somewhat seen as a conversion pass (even if an
inter-dialect conversion) as it translates any `pyHIR.func` operation to
`pyHIR.globalFunc`. Operations from the `py` dialect replace the `func`
op to create the resulting function object with the default arguments
together with the `pyHIR.globalFunc`.

One of the features not yet implemented are captured values. Since
`pyHIR.func` is not isolated from above it can use any MLIR values
defined outside. These will need to be lowered as "fields" of the
function object in the future, but is not yet done so as part of this
pass. Rather, it currently emits an error in this case.

Running this pass is required somewhere within in the pass-pipeline as
the `pyHIR` to `py` pass will likely only support `globalFunc`.