[CodeGenNew] Initial implementation reading and writing locals and globals
This implements the initial logic for name resolution. Local variables are immediately translate to SSA form using the `SSABuilder` while global variables are translated to reads and writes to the global dictionary.
The code only implements the happy-path. No exceptions are currently thrown on lookup failure, nor is the fallback to builtins implemented yet.