[pyHIR] Add `call` op
This op is designe to exactly mirror the way call expressions work in python but relaxes the syntax a bit to allow future optimizations such as inline expansions.

This PR only includes the dialect changes itself with frontend and lowering changes coming later. Given that elaborating on the semantics would require explaining python's call sequence in full, the documentation of the op has been left close to empty as well.