[PyHIRToPylirPy] Implement `pyHIR.call` lowering (#115)
In the general case, the lowering is implemented the same way as the old
frontend: The call is lowered to the `pylir__call__` intrinsic which
implements the method lookup. The implementation of `pylir__call__` is
additionally generated during the conversion pass and not in the
frontend like the old codegen. The implementation of `pylir__call__` was
copy-pasted for now. It is likely that it will be simplified in the
future by being able to use `pyHIR` operations. As soon as the new
`CodeGen` is done we can also delete the old one.