[CodeGenNew] Implement intrinsic calls (#119)
Intrinsic calls are a very important part of the builtins implementation
as they make it possible to directly create MLIR implementations
implementing semantics that are impossible in pure python code. The
kinds of intrinsics are all automatically generated using TableGen from
the `py` dialect.

This PR hooks these intrinsics up to the new codegen. The code is mostly
taken from the old but adjusted for the evolved style.