[PyHIRToPylirPy] Add exception handling lowering
This PR implements the lowering of all exception handling variants and creates the infrastructure for future ones. This is implemented by a custom pattern rewriter called `ExceptionRewriter` that automatically rewrites any `py` ops with an exception handling version to that version if lowering an exception hanling operation.
This allows using the exact same code for lowering patterns for the normal and exception handling version of an operation.