[CodeGenNew] Implement `for` statement (#132)
With `try` implemented, implementing `for` is simple. Calls to `next`
and `iter` are generated and stop iteration exceptions handled.

In the future we may want a higher-level op to efficiently lower and
represent iterations over builtins (tuples, lists, etc). This is left as
a future task to avoid premature optimization right now.