[CodeGenNew] Implement almost all comparison operators
Comparison operators are not just binary operators like in most languages but can be chained which implicitly acts like `and` expressions of comparison operators.
Additionally, `is` and `in` are implemented as comparison operators in python that can also be negated using `not`.
This PR implements all of them except `in` as there isn't yet a lowering in `pyHIR`.