[Parser] Verify body of `const_export` classes (#140)
Class bodies in python are normal code used to construct the namespace.
This is too permissive for the purpose of the `const_export` macro which
tries to create constant instances of the type.

Classes annotated with the intrinsic must therefore only contain very
few allowed statements and verified by the semantic analysis that this
is the case. This most importantly serves as an invariant for later
`CodeGen` phases.