[CodeGenNew] Implement simple `raise`
This PR implements the most common form of `raise` with a single expression. Depending on whether the expression evaluates to a type object or not an instance of an exception is constructed and subsequently raised via `py.raise`.
Implementing just `raise` as well as attaching the cause and context are left as future TODOs.