[CodeGen][Main] Refactor import logic (#149)
The previous import logic had an odd split of responsibilities where the
codegen was performing the I/O operations of finding the module and then
handed the file handle to the frontend invocation performing the module
load. The frontend would then return the module qualifier to use by
codegen for the import, creating an odd dependency and back-and-forth.

This PR therefore refactors this logic a bit and makes the callback used
by the codegen a pure "notifier" so-to say. The codegen constructs the
qualifier used to denote the module notifies the frontend of requiring
the module to be loaded. It uses the same qualifier to call the modules
initialization function. The frontend receives the qualifier and
performs the whole logic starting from finding the python source file to
then spawning the codegen task.

The threading logic was also simplified in the process by using an
explicit `llvm::ThreadPool` and a `llvm::ThreadPoolTaskGroup`.