[CodeGenNew] Partially implement `import <module>`
Like the old codegen, the import functionality triggers initialization of the module which for now is required for the builtin module. It does not yet implement initialization of the module object or any name binding. The logic is mostly taken from the old codegen with the expectation of it being removed very soon.