[PyHIR] Add `pyHIR.initModule` op (#151)
This operation is used to call the `init` op body of any module that is
not the main module and is required for the purpose of module import.
Right now it only calls the init method without 1) creating the module
object, 2) raising import errors or 3) ensuring the module hasn't been
initialized previously. These are left as future tasks