[CodeGenNew] Prevent infinite `bool` recursion (#164)
Any "truthy" evaluation in python unconditionally calls the `bool`
constructor in `CodeGenNew`. This is problematic as the implementation
of `bool` may also use `if` statements and similar, triggering "truthy"
evaluation and therefore an infinite recursion.

Fix this in the same way as in `CodeGen` by creating checks whether the
value being evaluated is already a `bool`, only calling `bool` if this
is not the case. The only con of this approach is higher complexity in
output-ir and potentially larger code-size. A future change may restrict
this kind of code-generation to only the builtin module.