[Py][CodeGenNew] Add `raiseEx` op for more uniform exception handling
The aim of having this op is to make exception handling more uniform and allow creating `raise` ops that jump to an exception handler within the same function.
Some code had to be adjusted as they were making assumptions about the operation that exception handling is being added to, not being a terminator.

A redesign of the `cloneWithExceptionHandling` API should be done in the future to make these differences (terminator vs non-terminator op) more transparent.