[PyHIR] Add required closure parameter to `globalFunc` (#170)
After outlining `globalFunc` must use the closure parameter to access
any non-local variables that were previously implicitly captured by the
`func`. The first step in implementing this is to add a closure
parameter to `globalFunc` to which the python `function` will be passed.
The lowering of `call` already performed the passing of the closure.
Only the CC version had to be adjusted to pass the closure to the
implementation.