[CodeGenNew] Implement `nonlocal` variables
`nonlocal` variables in Python are variables that are defined within a parent function or class and captured by the current class.
Like normal locals, the variables can be assigned to and the change will be visible to all functions that have access to the variable.

This is implemented using a so-called cell in Python. The function creating the variable that will get captured by sub-functions creates a cell for every captured value. The cell gets captured by sub-functions. Reads and writes of the variable are redirected to the single slot of the cell, creating a single reference everyone has access to.