[CodeGenNew] Correctly save and restore `finally` and `try` on function change
Prior to this PR the current exception handler and finally stack were not cleared when entering a new function. This leads to invalid code being generated referencing blocks and variables outside the current function.

This PR fixes that issue by introducing a proper `ChangeFunction` class that sets and resets all state associated with function generation.