[CodeGenNew] Implement iterable unpacking
The `py` dialect already contains an existing operation to perform this operation. We will probably want to move this operation to the `HIR` dialect in the future as that is the more fitting abstraction level. Until then, the implementation for iterable packing is the same as in the old CodeGen.