[CodeGenNew][PyHIR] Implement attribute references (#186)
Attribute references are the last kind of expressions missing before
being able to replace the old `CodeGen`. This PR implements the lowering
of attribute refs to HIR ops mirroring the semantics. Most notably,
accessing an attribute calls `__getattribute__`, not `__getattr__`.
These are then later lowered to the same intrinsic calls that the old
`CodeGen` lowers to.