[Main] Remove `-emit-mlir` (#188)
This compiler option is the most arbitrary of all the `-emit` options as
it outputs Pylir IR before it is lowered to LLVM IR. Since multiple
lowerings have happened at that point in time, the currently defined
pipelines will crash on the given operations. The name was arbitrary as
well.

Remove the option for now. A more composable pipeline that does input
detection (like IREE) can be introduced in the future with finer grained
output options to match.