[PyHIR] Prevent capturing of constant operations
Constant operations in MLIR are meant as a bridge between compile time constants, attributes, and SSA Value definitions.
As such, capturing the results of a constant operation during outlining is unnecessary and any such use should simply be replaced with a local copy of the constant operation within the resulting `globalFunc` operation.