[PyHIR] Add ops for reading and writing module attributes
These ops are the first step to introducing module objects into the frontend. Of note is that modules attributes are equal to the global dictionary of a module being initialized. These ops are therefore specially designed for this kind of access.

The lowering is currently far from perfect as it makes the assumption of a `__dict__` slot being present and containing a dictionary in the modules object. This creates an undesirable coupling to the frontend. I do not yet have a better idea how to implement this