[CodeGen] Implement lowering of `class`s to `pyHIR.class` (#199)
Class bodies in Python is normal python code except with different name
lookup rules. All name lookups are by default redirected to the
dictionary that later becomes the class body. If at runtime, a value is
not yet bound in the dictionary, the normal lookup procedure for
non-locals and globals continues.

The bulk of this PR is implementing this namelookup change and any other
required state changes for class definitions. A class definition can now
be lowered to a `pyHIR.class` op.