[Optimizer] Switch the `O0` pipeline to DCE
The pipeline previously used canonicalization which did optimization which 1) does not fit the purpose of the O0 pipeline (no optimization) and made us accidentally rely on a canonical form within lowering. One such case was found in the expand-py pass where it failed to legalize operations that usually did not occur due to canonicalizations.