[Optimizer] Disable region simplification in `canonicalizer` by default (#216)
Region simplifications in the canonicalizer refers to 1) dead block
elimination, 2) optimistic dead-value elimination and 3) identical
region collapsing. From all of these we only want 1) to run similarly as
often as pattern application which we can achieve by scheduling the DCE
pass. (note that we might need to run them together until a fixpoint is
reached). 3) has been observed to be detrimental in some cases and
expensive in compile time.